+++ packages/indexer-database/src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from "./BaseRepository";
+export * from "./BlockchainEventRepository";
